+++ voxygen/src/hud/mod.rs
@@ -40,7 +40,7 @@ pub use loot_scroller::LootMessage;
 pub use settings_window::ScaleChange;
 pub use subtitles::Subtitle;
 
-use bag::Bag;
+use bag::BagManager;
 use buffs::BuffsBar;
 use buttons::Buttons;
 use change_notification::{ChangeNotification, NotificationReason};
@@ -915,7 +915,7 @@ pub struct Show {
     intro: bool,
     crafting: bool,
     bag: bool,
-    bag_inv: bool,
+    bag_menu_split: bool,
     bag_details: bool,
     trade: bool,
     trade_details: bool,
@@ -956,7 +956,7 @@ impl Show {
             intro: false,
             crafting: false,
             bag: false,
-            bag_inv: false,
+            bag_menu_split: false,
             bag_details: false,
             trade: false,
             trade_details: false,
@@ -3533,10 +3533,10 @@ impl Hud {
                 poises.get(entity),
             )
         {
-            for event in Bag::new(
+            for event in BagManager::new(
+                global_state,
                 client,
                 &info,
-                global_state,
                 &self.imgs,
                 &self.item_imgs,
                 &self.fonts,
@@ -3561,9 +3561,7 @@ impl Hud {
             .set(self.ids.bag, ui_widgets)
             {
                 match event {
-                    bag::Event::BagExpand => self.show.bag_inv = !self.show.bag_inv,
-                    bag::Event::SetDetailsMode(mode) => self.show.bag_details = mode,
-                    bag::Event::Close => {
+                    bag::BagEvent::Close => {
                         self.show.stats = false;
                         Self::show_bag(&mut self.slot_manager, &mut self.show, false);
                         if !self.show.social {
@@ -3577,18 +3575,24 @@ impl Hud {
                             self.events.push(Event::TradeAction(TradeAction::Decline));
                         }
                     },
-                    bag::Event::ChangeInventorySortOrder(sort_order) => {
+                    bag::BagEvent::MoveBag(pos) => {
+                        global_state.settings.hud_position.bag.own = pos;
+                    },
+                    bag::BagEvent::BagExpand => {
+                        self.show.bag_menu_split = !self.show.bag_menu_split
+                    },
+                    bag::BagEvent::SetDetailsMode(mode) => self.show.bag_details = mode,
+                    bag::BagEvent::ChangeInventorySortOrder(sort_order) => {
                         self.events
                             .push(Event::SettingsChange(SettingsChange::Inventory(
                                 Inventory::ChangeSortOrder(sort_order),
                             )));
                     },
-                    bag::Event::SortInventory(sort_order) => {
+                    bag::BagEvent::SortInventory(sort_order) => {
                         self.events.push(Event::SortInventory(sort_order))
                     },
-                    bag::Event::SwapEquippedWeapons => self.events.push(Event::SwapEquippedWeapons),
-                    bag::Event::MoveBag(pos) => {
-                        global_state.settings.hud_position.bag.own = pos;
+                    bag::BagEvent::SwapEquippedWeapons => {
+                        self.events.push(Event::SwapEquippedWeapons)
                     },
                 }
             }